requirements.txt
fastapi
uvicorn
requests
pytz
python-dotenv
huggingface_hub
